fix(clipboard): move useCopyFeedback to src/lib per task acceptance criteria
PR #277 review noted the task's acceptance criteria specify
src/lib/useCopyFeedback.ts, but the hook landed at src/lib/hooks/useCopyFeedback.ts.
Move it to the specified path and update all 8 call sites + the unit test import.
