+++ src/components/image-convert/ImageConvertPage.tsx
@@ -13,7 +13,6 @@ import { useCallback, useEffect, useState } from 'react';
 import { FileDropzone } from '@/components/common/FileDropzone';
 import { Button } from '@/components/ui/button';
 import { useBatchProcessing } from '@/lib/hooks/useBatchProcessing';
-import { useCopyFeedback } from '@/lib/hooks/useCopyFeedback';
 import { useToolAnalytics } from '@/lib/hooks/useToolAnalytics';
 import { useToolSettings } from '@/lib/hooks/useToolSettings';
 import { createId, downloadBlob } from '@/lib/tools/image-common';
@@ -25,6 +24,7 @@ import {
 	validateImageFile,
 } from '@/lib/tools/image-convert';
 import { buildZip, dedupeZipNames } from '@/lib/tools/zip';
+import { useCopyFeedback } from '@/lib/useCopyFeedback';
 import {
 	ConvertOptionsPanel,
 	type ConvertUiOptions,
